linux-yocto: Smack config and patches
For any version of linux-yocto, we enable the Smack kernel config options.

For 3.14, we add one crucial patch which is required for booting with systemd.
Tizen Common has additional 3.14 kernel patches which fix issues and which
may have to be imported, too. This creates a dependency on oe-core with
a 3.14 .bb; at the moment, this layer gets tested with Poky Dizzy.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>