smack-userspace: initial recipe
This was originally copied from meta-tizen, then rewritten entirely as a
normal Bitbake recipe and updated to 1.1.0.

The recipe is intentionally not named just "smack", because that word will be
used as override controlling the Smack feature, which would cause conflicts
when compiling a recipe which has a "smack" package (for example, for
FILES_smack). Renaming the entire recipe is the simplest solution and also
distinguishes more clearly between a Smack feature of the system and the
specific userspace component.