systemd: compile as in Tizen
This adds all current (as in accepted/tizen/mobile/20150208.065204) Tizen
patches and enables Smack support in systemd 2.16 (as found in OE-core Dizzy).

Some Tizen patches conflict with OE's own patches: they seem
to be irrelevant for security, so they are just committed to git
for the sake of completeness without being used.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>