smack setup: define rules and set labels
The Tizen three-domain-model rules get created as part of the base-files.
They were taken from a live Tizen system, mirroring the rules as defined
in default-ac-domains.

Setting the corresponding Smack labels and types has to be split between
base-files (regular directories in the filesystem) and
systemd-tmpfiles-setup.service (for directories created by systemd).

Setting Smack labels is done by calling chsmack directly. This avoids
the dependency on a patched rpm with manifest support and is also
the direction into which Tizen is going.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>