layer: introduce smack override
Making the meta data changes depend on a new "smack" override has the
desirable effect that one can include the layer in a bblayers.conf and then
decide separately whether Smack support is really enabled.

The goal is to do that via DISTROOVERRIDES in a distro conf, potentially
depending on a distro feature.

The Smack userspace builds even when Smack support is disabled.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>