tizen-platform-wrapper: initial recipe
Needed for security-manager (through security-manager->tizen-platform-config->
tizen-platform-wrapper and
security-manager->libslp-db-util->dlog->tizen-platform-config->
tizen-platform-wrapper).

A "tizen" layer would be a better place for this. Maintaining it here for now.

Also needs to be built natively, because the tzplatform-tool will be invoked
during compile time by tizen-platform-config.

The not very helpful description of the recipe comes straight from
Tizen. Better explanations what this code does and how it differs
from tizen-platform-config welcome...

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>