ima-evm-utils: initial recipe
This mirrors the current state in Tizen. The Tizen git repo is
used because it seems to be more up-to-date than the SourceForge
repository (http://sourceforge.net/projects/linux-ima/).

The OpenSSL exception was added in upstream commit 2b89313
and documented in the header of each source file, so linking
against OpenSSL is okay.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>