meta-security-framework: add auditd and its libcap-ng dependency
Some (but not all) audit messages are also printed to syslog when
auditd is not installed. So only enable this for debugging, with
   CORE_IMAGE_EXTRA_INSTALL += "auditd"
in local.conf.

These recipes were copied verbatim from meta-selinux
2ebacbad21bbc66e0a4eba123bd7c8d6b14c8a27 because adding that
layer also enables SELinux.

The recipes get added to meta-security-framework because
like the kernel change itself, they are of use independently
of the active security mechanisms in the kernel.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>