meta-security-framework: support audit logging
The audit kernel subsystem is useful as debugging tool. However,
there are concerns about its performance in particular in
combination with the default audit.conf. See
http://thread.gmane.org/gmane.comp.emulators.kvm.devel/135223
https://fedorahosted.org/fesco/ticket/1311

So only enable this for debugging, with
   SRC_URI_append_pn-linux-yocto = " file://audit.cfg"
in local.conf.

Some (but not all) audit messages are also printed to syslog when
auditd (to be added separately) is not installed.

The recipe gets added to meta-security-framework because it is
of use in several different scenarios (with or without Smack,
with or without IMA).

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>