systemd 2.19: fix IMA policy loading
This reverts broken changes that went into systemd 2.18. The IMA
developers have been notified.

The patch gets added to the systemd_219.bbappend in meta-security-smack
to minimize the number of places where systemd gets modified. Depending
on how the IMA policy is going to be loaded, this creates an indirect
dependency from meta-integrity to meta-security-smack which will be
documented separately.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>