dbus: enable Cynara privilege checking
The patches add a new <check> rule which calls Cynara while processing
D-Bus messages to determine whether sending resp. receiving the
message is allowed. All of these patches are non-upstream and most
likely never will be, because the depend on adding support for
non-blocking privilege checking in the core of the D-Bus daemon, which
upstream is unwilling to accept.

At the moment, the Samsung security team in Poland maintains these
patches for Tizen. The port to D-Bus 1.8.10 was done for this recipe
and needs to be cleaned up in coordination with the Samsung maintainers,
see https://bugs.tizen.org/jira/browse/TC-2520

Calling Cynara from the dbus-daemon leads to a cyclic dependency (dbus
-> cynara -> systemd -> dbus), something that Bitbake (in contrast to
OBS) does not support. To break the cycle, the regular "dbus" recipe
gets modified such that it only provides the D-Bus libraries, but not
the daemon. The daemon then gets compiled separately in the
"dbus-cynara" recipe.

As discussed and agreed in
https://review.tizen.org/gerrit/#/c/31310/4/bus/system.conf.in the
0012-Perform-Cynara-runtime-policy-checks-by-default.patch which
modifies the default bus policies is not final either: send_method
should stay forbidden by default. Right now, the latest patch got
copied from Tizen unmodified.

In a nutshell, the modified bus policies limit what processes can do
on the user session bus: if they have the
http://tizen.org/privilege/internal/dbus privilege, nothing changes,
but if they do not, they cannot communicate with services unless
additional rules for that service allow that. The system bus already
was locked down like that, so the modification merely restricts access
to broadcast signals a bit more.

Because this is a potentially intrusive change that might not be
desired by all users of the layer, dbus-cynara must be enabled
explicitly by adding "dbus-cynara" to DISTRO_FEATURES. The necessary
Cyanara rules granting "System" and "User" process the
http://tizen.org/privilege/internal/dbus privilege can be created
by installing the "security-manager-policy" package.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>