ima-evm-utils: support compilation on older Linux distros
For example, Ubuntu 12.04 does not yet have the IMA defines in
linux/xattr.h. This affects us when building ima-evm-utils-native,
because that uses the headers from the host (in contrast to
compilation for the target, which uses a recent enough
linux-libc-headers).

Upstream source code gets patched to be more permissive.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>