keyutils: avoid QA warning about request-key-debug.sh
By removing the file from the staging, we avoid the following warning:
QA Issue: keyutils: Files/directories were installed but not shipped
  /usr/share/request-key-debug.sh [installed-vs-shipped]

This looks like a debugging script which is not really needed in
the image, so this looks like the simples solution. Adding it to
the keyutils-dbg pacckage would also work, but only if those
get generated, which is not guaranteed.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>