meta-security-framework dbus: avoid hard dependency on D-Bus 1.8.10
OE-core master has updated D-Bus to version 1.8.18. This breaks
compilation when our layer is enabled, because each .bbappend
file must have a matching .bb file.

Our .bbappend file works regardless of the base dbus recipe's version,
so we can use a wildcard instead of a specific version to make our
layer compatible with both OE-core Fido and master at the same time.

The patched dbus-cynara remains at version 1.8.10 for now. This is not
a problem (libdbus 1.8.18 is compatible with dbus-daemon 1.8.10) but
probably dbus-cynara should be updated to 1.8.18 to take advantage of
upstream fixes.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>