xattr-images.bbclass: preserve xattr in .jffs2 and .tar.bz2
OE-core does not package xattrs in .jffs2 and .tar.bz2 images by
default. When building such images with Smack and/or IMA/EVM image
signing enabled, this class needs to be inherited to ensure that
xattrs are not getting lost.

Beware that "--xattrs --xattrs-include=*" and GNU tar >= 1.27 are
required when extracting xattrs from such an image.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>