coreutils: do not accidentally link against libsmack
The coreutils configure uses libsmack if installed, which is
unintentional and non-deterministic. If we compile with Smack
disabled, we need to disable the feature explicitly.

Found when compiling with a sysroot which still had libsmack from a
previous compile run with a different local.conf. Should not happen
when compiling from scratch.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>