ima-evm-rootfs: support hashing of files
It may be useful to hash some files (= record just their hashsum in
security.ima) instead of signing them (= stored the hashsum signed
with the private key) because hashed files then can be modified on
the target.

The default does not change: everything is signed.

To sign or hash everything, use something like this:
   WRITABLE_FILES = "-path './etc/*' -o -path './var/*'"
   IMA_EVM_ROOTFS_SIGNED = ". -type f -a ! \( ${WRITABLE_FILES} \)"
   IMA_EVM_ROOTFS_HASHED = ". -type f -a \( ${WRITABLE_FILES} \)"

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>