ima-evm-rootfs: add iversion to fstab
If requested (default is off), the fstab entry of certain file systems
gets changed such that they get mounted with "iversion" (aka
"i_version" in busybox). This is required for file systems and
policies that allow writing, because otherwise IMA will not notice
file modifications.

To enable the feature, use a space-separated list of mount points:
   IMA_EVM_ROOTFS_IVERSION = "/ /data"

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>