ima_policy_hashed: sign or hash everything
This is the first policy which actually could be used in practice. It
requires that all files are either signed or hashed. Signed files are
read-only whereas hashed files are read/write (their hash will get
updated on modifications). New files also get hashed automatically.

At the moment, it still does not offer full protection. Online attacks
are possible by writing a new file and replacing a read-only file with
it. Offline attacks are possible by modifying a file and hashing it.

To address these kind of attacks, additional changes are needed (like
mounting the rootfs read-only and enabling EVM).

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>