IMA: simplify kernel configuration
The original ima-evm.cfg was copied from Tizen and did not match the
upstream kernel. The new ima.cfg is much closer to the upstream
recommendation.

In addition, it no longer depends on X.509 key loading from the
kernel, which turned out to have a security problem (patch pending
upstream). Instead, the initramfs now also loads the key.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>