app-runas: test tool for running commands as applications
The "app-runas" command uses the SecurityManager API to register an
app for a user (with appid, pkgid and uid chosen via command line
parameters) and then runs arbitrary commands as that app.

Intended usage is manual and automated testing. Therefore the command
gets both packaged (for use in local.conf with
"CORE_IMAGE_EXTRA_INSTALL += app-runas") and deployed (for "testimage"
testing).

Example:
  adduser -D tester
  app-runas -a test-appid -p test-pkgid -u `id -u tester` -i
  app-runas -a test-appid -e -- bash

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>