packagegroup-security-framework.bb: Cynara, SecurityManager
Useful because it moves the definition out of image recipes and makes
it available also for on-device installation via package feeds.

These components only make sense when used together. Users of this
package group must be aware that the functionality of all current
components in this group depend on Smack.

It is not disabled altogether without Smack (at least at the moment),
because perhaps at some point there will be components which do not
depend on Smack.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>