smack testing: remove VERBATIM licensed code
The test program comes from the mmap man page, which has a
non-standard license. The license is not fully understood at this time
and may impact recipients trying to use the layer with just the
normal, well-known licenses. Therefore the code gets removed while
investigating the situation.

Instead, a non-functional replacement is used, which keeps the recipe
and everything that depends on the presence of a "mmap_test" binary
working. This is less intrusive than removing the entire recipe. Once it
is clear that the removal is permanent, we can also remove the recipe.

The replacement is MIT licensed again.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>