smack tests: push test binaries to /tmp instead of /usr/bin
Changed the test binaries copyTo path to /tmp/ from /usr/bin.
Adapted affected scripts to look for test binaries in PATH, and
if they are not found, check in /tmp/, or fail otherwise.
This is needed in order to differentiate between test and default
binaries, which are protected by IMA signature.