securitymanager.py: fix test on distros with default policy
The default policy has no privileges which map to Unix groups and thus
we get an empty list of privileges when querying the database. However,
that the query runs already confirms that SecurityManager was set up, so that is
good enough for the very basic test.

The other tests can fall back to some random "FoobarPrivilege" because
SecurityManager itself does not (and cannot) check against a list of
known privileges.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>