TravisCI: minimize disk usage
For initial compile runs with empty sstate cache the limited disk
space (around 10GB) is typically the reason why jobs abort, even
before hitting the runtime limit. Therefore it is important to make
efficient use of the available disk space.

rm_work.bbclass turned out to schedule tasks sub-optimally, because it
switches from building one recipe foo to building a more important
recipe bar which depends on foo without finishing (and cleaning) the
started build of foo. Grouping tasks of one kind and ordering them so
that completion and cleaning is more important than starting a new
build avoids this problem.

The current implementation is somewhat fragile because it hard-codes
the list of tasks. Adding unknown ones breaks it, because those then
end up with their original priority as determined by the
recipe. Detecting all tasks and their topological ordering at runtime
would be better.

This reordering is implemented in a custom scheduler. That scheduler
also implements an optional additional limit on the number of parallel
compile tasks. This turned out to be useful on the TravisCI Trusty
environment, because starting to compile complex recipes like
qemu-native and linux-yocto in parallel had the effect that none of
them completed before the runtime limit, which repeated over and over
when restarting the job.

Besides ensuring that do_rm_work actually runs sooner, it also gets
extended by rm_work_and_downloads.bbclass such that downloaded sources
also get removed. This can only be done reliably when it is guaranteed
that no other recipe needs the same download, otherwise removing the
downloaded file breaks the other recipes when their do_fetch has
already run and is considered completed. To make removal reliable,
each recipe gets its own private DL_DIR. The downside of course is
that the same source will get downloaded multiple times during a
complete build, but that is a reasonable tradeoff in an environment
with high network bandwidth and limited disk space.

Because do_fetch tasks do not get rerun after removing a download,
rm_work_and_downloads.bbclass cannot be used for incremental
builds. Again, that is not an issue for builds inside TravisCI, which
always start with a clean build directory.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>