dbus-oe-core.inc: 1:1 copy of dbus.inc from OE-core
OE-core master recently [1] removed dbus.inc. Including the versioned
recipe instead would introduce a tighter coupling wth OE-core and not
work with older branches, so we have to copy the build rules under a
name that does not conflict with the original name on branches which
still have it.

This is a 1:1 copy of dbus.inc e26e348d7b6301f0e1cb15018ebadfa0eea4013c.

[1] http://cgit.openembedded.org/openembedded-core/commit/?id=aea82066eb3ba67c1d3a170112ff3bb324b43dfb

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>