dbus-cynara: use copy of dbus.inc
OE-core master recently [1] removed dbus.inc. Including the versioned
recipe instead would introduce a tighter coupling wth OE-core and not
work with older branches, so we have to use a copy of the build rules
under a name that does not conflict with the original name on branches
which still have it.

This also allows removing the workaround which was added earlier for
changes made to dbus.inc on the fido and dizzy branches.

[1] http://cgit.openembedded.org/openembedded-core/commit/?id=aea82066eb3ba67c1d3a170112ff3bb324b43dfb

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>