rmwork.py: avoid hard-coding task list
Instead, look at the tasks of each recipe as serialized by our base
class and merge these task lists into one larger list. This works
because recipe tasks lists only differ by sometimes having additional
tasks, but not in the order of the common ones.

Both the speed and the completion scheduler work as base.

Debug output is intentionally left in the code, but commented out.
This way, it is easier to enable it again during debugging.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>