rmwork.py: use speed scheduler as base
The speed scheduler has the advantage that it priotizes recipes that
other recipes depend on first. This looks like a useful property even
when reordering priorities again such that there is no massive
parallelism.

It also paves the way for removing the completion scheduler altogether
and replacing it with the rmwork.py scheduler.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>