smack: fix transmuting Smack directories in rootfs
Setting the Smack transmute flag on /etc only had an effect on real
devices. It has no effect on files added to /etc during image creation,
because then the special flag is handled by pseudo, which does not know
about the special semantic.

The right fix would be to enhance pseudo. But that is quite a bit of
work for very little gain, so instead an easier workaround is used: all
images which use Smack should inherit xattr-images.bbclass, so that class
gets extended to find and fix transmuting directories.

In addition, the same code could also sets the _ label explicitly on
all files which would have no label otherwise. This is not needed for
running the rootfs on a device with Smack enabled, and thus not done yet,
in particular because it leads to a lot more xattrs, which confuses
mkfs.ext4 (https://bugzilla.yoctoproject.org/show_bug.cgi?id=8992).

It might be useful for build tools which need to see the final xattrs
already in the rootfs. One such tool is swupd: it unpacks files on the
target with tar, and while GNU tar knows that it needs to set the
security.SMACK64="_" xattr if that was in the archive, it does not
know enough about Smack to remove the default Smack label inherited
from the process (like "System") when the file has no such xattr in
the archive.

The test for Smack labels in a running system also gets updated to
cover at least one file which is expected under /etc (/etc/passwd) and
should have the same Smack label as /etc (System::Shared). Picking a
suitable file is tricky, because in a stateless distro, /etc might
actually be empty. Right now, that is not supported by the test.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>