smack: fix transmuting /etc in rootfs, explicit _ label
Setting the Smack transmute flag on /etc only had an effect on real
devices. It has no effect on files added to /etc during image creation,
because then the special flag is handled by pseudo, which does not know
about the special semantic.

The right fix would be to enhance pseudo. But that is quite a bit of
work for very little gain, so instead an easier workaround is used: all
images which use Smack should inherit xattr-images.bbclass, so that class
gets extended to find and fix transmuting directories.

In addition, the same code also sets the _ label explicitly on all files
which would have no label otherwise. This is not needed for running the
rootfs on a device with Smack enabled. It gets done because it helps tools
which look at xattrs without understanding Smack to determine what the
Smack label will be resp. has to be.

One such tool is swupd: it unpacks files on the target with tar, and
while GNU tar knows that it needs to set the security.SMACK64="_"
xattr if that was in the archive, it does not know enough about Smack
to remove the default Smack label inherited from the process (like
"System") when the file has no such xattr in the archive. So as a
temporary workaround for use of swupd with unpatched tar, all files
get an explicit security.SMACK64 to avoid the problem.

The test for Smack labels in a running system also gets updated to
cover at least one file which is expected under /etc (/etc/passwd) and
should have the same Smack label as /etc (System::Shared). Picking a
suitable file is tricky, because in a stateless distro, /etc might
actually be empty. Right now, that is not supported by the test.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>

merge: test