xattr-images.bbclass: avoid python-xattr dependency
python-xattr is a non-standard module which we cannot assume to be
installed. Building it also is not easy, because it would imply
building python-native first (doing it with Python already installed
on the host would depend on its development files, which also cannot
be assumed to be installed). Going back to pure shell script would
imply calling chsmack several times for each filesystem entry in the
rootfs, which is expected to be much slower than just calling a native
C function.

That leaves re-implementing xattr with ctypes, which is a standard
Python module.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>