base-files_%.bbappend: /var writable for system services
/var is like /etc in the sense that different system services need
the ability to write there. We could do this so that each service
must label its own files, but that's an enhancement that can come
later. For now, configure Smack exactly as for /etc.

Together with the commit that fixes transmute semantics, this change
fixes a problem where /var/lib/connman/wired.config was labeled with
"_", which prevented /usr/share/connman/wired-setup from updating the
config file. The effect was that the IPv4 address passed via
boot parameters where not used when running under runqemu.

Creating directories in the target filesystem is better done with
"install -d" than "mkdir -p" (because of permissions), so use that and
also fix the old code for /etc.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>