travis: support compilation outside of Travis CI
After moving most of the embedded setup commands from .travis.yml into
travis-setup.sh and tweaking them to be less dependent on the specific
TravisCI environment, compiling on a local machine becomes easier. The
goal is to pre-populate the S3 sstate cache from a fast build machine
and then run normal, automatic testing again in TravisCI.

sstate2s3.sh no longer modifies the local sstate cache and instead
excludes files via an exclude file. That ensures that the cache can be
re-used between invocations on a local machine and it may also allow
uploading of the sstate files in parallel to the actual compilation.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>