e2fsprogs_git.bb: skip recipe when OE-core is recent enough
This recipe is only meant as replacement for older OE-core releases.
When building against current OE-core master, we want to use the
upstream version.

Testing against the known e2fsprogs_1.42.9.bb works because that has
been in OE-core for a long time. Testing for e2fsprogs_git.bb would
also work, but could be ambiguous.

We need to skip the recipe explicitly because currently the versions
are identical. Even if we lowered the one used here, layer priorities
still cause this one to be picked.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>