linux-%.bbappend: Add IMA patch fixing unneeded re-appraisal
The patch taken from https://sourceforge.net/p/linux-ima/mailman/message/34894743/
makes IMA not to re-appraise a new file if only its timestamps
have changed during extraction from a tarball.

This is needed to avoid resetting the security.ima attribute for files
extracted from tarballs with software updates.

Signed-off-by: Dmitry Rozhkov <dmitry.rozhkov@linux.intel.com>

linux: source code patch version dependent, refreshed

Applying source code patches is risky and can easily break when the
upstream source changes. To avoid this, the patch only gets applied to
specific kernel source versions that are known to work. The downside
of course is that the patch will not get applied to future versions
where it is still needed; this will have to be caught by testing.

Apparently the patch in the previous commit had some whitespace issues,
at least it did not apply and had to be re-generated.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>