xattr-images.bbclass: avoid executing swupd
Embedding a Python script inside a shell wrapper is unnecessary
(IMAGE_PREPROCESS_COMMAND can invoke Python code functions directly,
and the initial cd can be replaced with the "dirs" function attribute)
and had unintended side effects: when a recent commit introduced
`swupd verify` in a comment, the shell wrapper called that and
injected its output into the Python script.

That merely led to some error output if swupd wasn't found, but when
installed on the host, it's embedded output broke the Python script
because of syntax errors.

The solution is to make xattr_images_fix_transmute a Python function.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>