smack: remove "User" domain
The entire "User" domain support has been tentative. It depends a lot
on the actual application framework and thus is neither necessarily
the same everywhere nor absolutely needed. For example, Ostro OS does
not need it.

Support for it also was incomplete, probably even in Tizen, leading to
problems even when a distro has no use for the "User" domain: in Ostro
OS, the PAM plugin sets the process label to "User" even when logging
in as root, and then when "passwd" is used to modify /etc/passwd, the
new file has "User" as Smack label because there is no transmute Smack
rule for User -> System::Shared and thus the transmute flag on /etc
gets ignored.

Instead of trying to fix this, it is easier to simplify the entire
layers by removing the "User" labels and rules. What remains is
hard-coded in the upstream recipes.

If someone need support for a "User" domain, they can add that to
their custom distro together with the code that depends on it or (when
that is not easy, for example in the smacknet tool) propose patches
which add back what is needed in a way that a) it is properly
documented and b) optional.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>