base-files_%.bbappend: punch SMACK-holes for /proc net interface entries.
Set SMACK labels to '*' for /proc/sys/net/ipv[46]/conf/*/*. These
procfs entries correspond to configurable settings of the network
interfaces. Setting the SMACK label to '*' allows access to these
at the SMACK level, letting the traditional DAC and capability
checks in the kernel take care of access control. Without this
systemd-networkd is unable to control packet forwarding of the
interfaces it otherwise manages (probably ditto for ConnMan).

The tmpfiles.d conf entry takes care of properly labeling entries
which do not correspond to network interfaces (all and default).
The udev rules take care of labeling real and virtual network
interfaces. Unfortunately it looks like we need two rules: one for
interfaces which get renamed by systemd and another for the ones
which don't.

Signed-off-by: Krisztian Litkey <krisztian.litkey@intel.com>