smack: enable moving Smack label loading to initramfs.
The Smack policy needs to be loaded before the IMA policy, if the Smack
labels need to be referenced in the IMA policy. If IMA policy is loaded
in initramfs, also the Smack policy needs to be loaded there.  However,
if IMA policy is loaded by (for example) systemd from rootfs, then the
Smack label files can be where they are.

Signed-off-by: Ismo Puustinen <ismo.puustinen@intel.com>